Add files via upload
This file gives our Macaulay2 code for applying the Hanlon-Hicks-Lazarev resolution of the diagonal to a polytope (F.4D.0047) from the polymake database, and checking for the existence of an ordering of the line bundles which appear on one side of the box product which makes the collection exceptional. This is a positive example because there does exist such an ordering.