Add files via upload
This file gives Macaulay2 code to enter the "Facets" polytope P, for polytope F.4D.0000 from the polymake database, apply the Hanlon-Hicks-Lazarev resolution of the diagonal to Y=X*X for X the toric variety associated to the toric polytope P, and construct the directed graph of all Hom's between line bundles appearing on one side of the external tensor product. The presence of a directed cycle in this directed graph (of length 2, for instance) is an obstruction to the possibility of an ordering of this collection of line bundles for which it is exceptional.